build: disallow building shared library
In the simplest case we either build/link everything (this project,
Cucumber-Cpp and Boost) static, or build/link everything shared.

If static then we don't have a problem.

If shared then we hit the issue that Cucumber-Cpp doesn't export all
required symbols by this project (i.e. CukeEngine).

So a more complex option might be to build/link Cucumber-Cpp statically
into the shared library of this project. But Cucumber-Cpp is a public
dependency, so we'd have to re-export all required symbols, and either
bundle or depend (only) on its public includes.

Further, if Boost is statically linked, and this project and
Cucumber-Cpp are both shared libraries, we get into ODR violation
territory.

But at least if we keep this project and Cucumber-Cpp static, then
Boost is free to be either static or shared.